+++ libmdtrie/bench/parser.hpp
@@ -13,11 +13,11 @@
 using namespace std;
 
 // Parse one line from TPC-H file.
-std::vector<int32_t>
+std::vector<uint64_t>
 parse_line_tpch(std::string line)
 {
 
-  vector<int32_t> point(TPCH_DIMENSION, 0);
+  vector<uint64_t> point(TPCH_DIMENSION, 0);
   int index = -1;
   bool primary_key = true;
   std::string delim = ",";
@@ -38,11 +38,11 @@ parse_line_tpch(std::string line)
       continue;
     }
     index++;
-    point[index] = static_cast<int32_t>(std::stoul(substr));
+    point[index] = static_cast<uint64_t>(std::stoull(substr));
   }
   index++;
   std::string substr = line.substr(start, end - start);
-  point[index] = static_cast<int32_t>(std::stoul(substr));
+  point[index] = static_cast<uint64_t>(std::stoull(substr));
 
   for (int i = 0; i < TPCH_DIMENSION; i++)
   {
@@ -56,11 +56,11 @@ parse_line_tpch(std::string line)
 }
 
 // Parse one line from TPC-H file.
-std::vector<int32_t>
+std::vector<uint64_t>
 parse_line_github(std::string line)
 {
 
-  vector<int32_t> point(GITHUB_DIMENSION, 0);
+  vector<uint64_t> point(GITHUB_DIMENSION, 0);
   int index = -1;
   bool primary_key = true;
   std::string delim = ",";
@@ -81,11 +81,11 @@ parse_line_github(std::string line)
       continue;
     }
     index++;
-    point[index] = static_cast<int32_t>(std::stoul(substr));
+    point[index] = static_cast<uint64_t>(std::stoull(substr));
   }
   index++;
   std::string substr = line.substr(start, end - start);
-  point[index] = static_cast<int32_t>(std::stoul(substr));
+  point[index] = static_cast<uint64_t>(std::stoull(substr));
 
   for (int i = 0; i < GITHUB_DIMENSION; i++)
   {
@@ -98,11 +98,11 @@ parse_line_github(std::string line)
   return point;
 }
 
-std::vector<int32_t>
+std::vector<uint64_t>
 parse_line_nyc(std::string line)
 {
 
-  vector<int32_t> point(NYC_DIMENSION, 0);
+  vector<uint64_t> point(NYC_DIMENSION, 0);
   bool primary_key = true;
   std::string delim = ",";
   auto start = 0U;
@@ -130,17 +130,17 @@ parse_line_nyc(std::string line)
     if (real_index >= 2 && real_index <= 5)
     {
       float num_float = std::stof(substr);
-      point[real_index] = static_cast<int32_t>(num_float * 10);
+      point[real_index] = static_cast<uint64_t>(num_float * 10);
     }
     else
     {
-      point[real_index] = static_cast<int32_t>(std::stoul(substr));
+      point[real_index] = static_cast<uint64_t>(std::stoull(substr));
     }
   }
 
   real_index++;
   std::string substr = line.substr(start, end - start);
-  point[real_index] = static_cast<int32_t>(std::stoul(substr));
+  point[real_index] = static_cast<uint64_t>(std::stoull(substr));
 
   point[0] -= 20090000;
   point[1] -= 19700000;
@@ -148,8 +148,8 @@ parse_line_nyc(std::string line)
   return point;
 }
 
-void update_range_search_range_tpch(std::vector<int32_t> &start_range,
-                                    std::vector<int32_t> &end_range,
+void update_range_search_range_tpch(std::vector<uint64_t> &start_range,
+                                    std::vector<uint64_t> &end_range,
                                     std::string line)
 {
 
@@ -168,13 +168,13 @@ void update_range_search_range_tpch(std::vector<int32_t> &start_range,
 
     if (start_range_str != "-1")
     {
-      start_range[static_cast<int32_t>(std::stoul(index_str))] =
-          static_cast<int32_t>(std::stoul(start_range_str));
+      start_range[static_cast<uint64_t>(std::stoull(index_str))] =
+          static_cast<uint64_t>(std::stoull(start_range_str));
     }
     if (end_range_str != "-1")
     {
-      end_range[static_cast<int32_t>(std::stoul(index_str))] =
-          static_cast<int32_t>(std::stoul(end_range_str));
+      end_range[static_cast<uint64_t>(std::stoull(index_str))] =
+          static_cast<uint64_t>(std::stoull(end_range_str));
     }
   }
   for (unsigned int i = 0; i < start_range.size(); i++)
@@ -187,8 +187,8 @@ void update_range_search_range_tpch(std::vector<int32_t> &start_range,
   }
 }
 
-void update_range_search_range_nyc(std::vector<int32_t> &start_range,
-                                   std::vector<int32_t> &end_range,
+void update_range_search_range_nyc(std::vector<uint64_t> &start_range,
+                                   std::vector<uint64_t> &end_range,
                                    std::string line)
 {
 
@@ -205,31 +205,31 @@ void update_range_search_range_nyc(std::vector<int32_t> &start_range,
     std::string end_range_str;
     std::getline(ss, end_range_str, ',');
 
-    int index = std::stoul(index_str);
+    int index = std::stoull(index_str);
 
     if (start_range_str != "-1")
     {
       if (index >= 2 && index <= 5)
       {
         float num_float = std::stof(start_range_str);
-        start_range[static_cast<int32_t>(index)] =
-            static_cast<int32_t>(num_float * 10);
+        start_range[static_cast<uint64_t>(index)] =
+            static_cast<uint64_t>(num_float * 10);
       }
       else
-        start_range[static_cast<int32_t>(index)] =
-            static_cast<int32_t>(std::stoul(start_range_str));
+        start_range[static_cast<uint64_t>(index)] =
+            static_cast<uint64_t>(std::stoull(start_range_str));
     }
     if (end_range_str != "-1")
     {
       if (index >= 2 && index <= 5)
       {
         float num_float = std::stof(end_range_str);
-        end_range[static_cast<int32_t>(index)] =
-            static_cast<int32_t>(num_float * 10);
+        end_range[static_cast<uint64_t>(index)] =
+            static_cast<uint64_t>(num_float * 10);
       }
       else
-        end_range[static_cast<int32_t>(index)] =
-            static_cast<int32_t>(std::stoul(end_range_str));
+        end_range[static_cast<uint64_t>(index)] =
+            static_cast<uint64_t>(std::stoull(end_range_str));
     }
   }
   start_range[0] -= 20090000;
@@ -238,8 +238,8 @@ void update_range_search_range_nyc(std::vector<int32_t> &start_range,
   end_range[1] -= 19700000;
 }
 
-void update_range_search_range_github(std::vector<int32_t> &start_range,
-                                      std::vector<int32_t> &end_range,
+void update_range_search_range_github(std::vector<uint64_t> &start_range,
+                                      std::vector<uint64_t> &end_range,
                                       std::string line)
 {
 
@@ -256,19 +256,19 @@ void update_range_search_range_github(std::vector<int32_t> &start_range,
     std::string end_range_str;
     std::getline(ss, end_range_str, ',');
 
-    int index = std::stoul(index_str);
+    int index = std::stoull(index_str);
     if (index > 10)
       index -= 3;
 
     if (start_range_str != "-1")
     {
-      start_range[static_cast<int32_t>(index)] =
-          static_cast<int32_t>(std::stoul(start_range_str));
+      start_range[static_cast<uint64_t>(index)] =
+          static_cast<uint64_t>(std::stoull(start_range_str));
     }
     if (end_range_str != "-1")
     {
-      end_range[static_cast<int32_t>(index)] =
-          static_cast<int32_t>(std::stoul(end_range_str));
+      end_range[static_cast<uint64_t>(index)] =
+          static_cast<uint64_t>(std::stoull(end_range_str));
     }
   }
